+++ workalendar/usa/massachusetts.py
@@ -10,3 +10,11 @@ from ..registry import iso_register
 class Massachusetts(UnitedStates):
     """Massachusetts"""
     include_patriots_day = True
+
+
+class SuffolkCountyMassachusetts(Massachusetts):
+    """Suffolk County, Massachusetts"""
+    FIXED_HOLIDAYS = UnitedStates.FIXED_HOLIDAYS + (
+        (3, 17, 'Evacuation Day'),
+        (6, 17, 'Bunker Hill Day'),
+    )